requirements.txt
torch
torchvision
doctr
opencv-python
pillow
pandas
openpyxl
customtkinter
ttkbootstrap
llama-cpp-python
matplotlib
defusedxml
deskew
pypdf
